requirements.txt
streamlit
flask
openai
PyYAML
requests